move 'binary' tests to separate file
